# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: BusinessNow PSA Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tasks:
    get:
      operationId: listTasks
      tags: [tasks]
      summary: List tasks
      parameters:
        - name: projectId
          in: query
          schema:
            type: integer
        - name: assigneeId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
    post:
      operationId: createTask
      tags: [tasks]
      summary: Create a task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTaskBody"
      responses:
        "201":
          description: Created task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
  /tasks/{id}:
    get:
      operationId: getTask
      tags: [tasks]
      summary: Get task by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
    patch:
      operationId: updateTask
      tags: [tasks]
      summary: Update task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTaskBody"
      responses:
        "200":
          description: Updated task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
    delete:
      operationId: deleteTask
      tags: [tasks]
      summary: Delete task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── USERS ────────────────────────────────────────────────────
...
  /time-entries:
    get:
      operationId: listTimeEntries
      tags: [timetracking]
      summary: List time entries
      parameters:
        - name: projectId
          in: query
          schema:
            type: integer
        - name: userId
          in: query
          schema:
            type: integer
        - name: startDate
          in: query
          schema:
            type: string
        - name: endDate
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of time entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TimeEntry"
    post:
      operationId: createTimeEntry
      tags: [timetracking]
      summary: Log a time entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTimeEntryBody"
      responses:
        "201":
          description: Created time entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TimeEntry"
  /time-entries/{id}:
    patch:
      operationId: updateTimeEntry
      tags: [timetracking]
      summary: Update a time entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTimeEntryBody"
      responses:
        "200":
          description: Updated time entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TimeEntry"
    delete:
      operationId: deleteTimeEntry
      tags: [timetracking]
      summary: Delete a time entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /timesheets:
    get:
      operationId: listTimesheets
      tags: [timesheets]
      summary: List timesheets
      parameters:
        - name: userId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
        - name: weekStart
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of timesheets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Timesheet"
    post:
      operationId: createTimesheet
      tags: [timesheets]
      summary: Create or get timesheet for a week
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTimesheetBody"
      responses:
        "201":
          description: Created timesheet
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Timesheet"
  /timesheets/{id}:
    get:
      operationId: getTimesheet
      tags: [timesheets]
      summary: Get timesheet by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Timesheet
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Timesheet"
        "404":
          description: Not found
    patch:
      operationId: updateTimesheet
      tags: [timesheets]
      summary: Update timesheet
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTimesheetBody"
      responses:
        "200":
          description: Updated timesheet
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Timesheet"
  /timesheets/{id}/submit:
    post:
      operationId: submitTimesheet
      tags: [timesheets]
      summary: Submit timesheet for approval
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Submitted timesheet
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Timesheet"
  /timesheets/{id}/approve:
    post:
      operationId: approveTimesheet
      tags: [timesheets]
      summary: Approve a submitted timesheet
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                approvedByUserId:
                  type: integer
      responses:
        "200":
          description: Approved timesheet
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Timesheet"
  /timesheets/{id}/reject:
    post:
      operationId: rejectTimesheet
      tags: [timesheets]
      summary: Reject a submitted timesheet
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                rejectionNote:
                  type: string
                rejectedByUserId:
                  type: integer
              required:
                - rejectionNote
      responses:
        "200":
          description: Rejected timesheet
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Timesheet"

  # ─── INVOICE LINE ITEMS ────────────────────────────────────────
...
components:
  schemas:
...
    Task:
      type: object
      properties:
        id:
          type: integer
        projectId:
          type: integer
        phaseId:
          type: ["integer", "null"]
        name:
          type: string
        status:
          type: string
        priority:
          type: string
        assigneeIds:
          type: array
          items:
            type: integer
        startDate:
          type: ["string", "null"]
        dueDate:
          type: ["string", "null"]
        effort:
          type: number
        billable:
          type: boolean
        isMilestone:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - projectId
        - name
        - status
        - priority
        - assigneeIds
        - effort
        - billable
        - isMilestone
        - createdAt
    CreateTaskBody:
      type: object
      properties:
        projectId:
          type: integer
        phaseId:
          type: integer
        name:
          type: string
        status:
          type: string
        priority:
          type: string
        assigneeIds:
          type: array
          items:
            type: integer
        startDate:
          type: string
        dueDate:
          type: string
        effort:
          type: number
        billable:
          type: boolean
        isMilestone:
          type: boolean
      required:
        - projectId
        - name
        - status
        - priority
        - effort
        - billable
    UpdateTaskBody:
      type: object
      properties:
        phaseId:
          type: integer
        name:
          type: string
        status:
          type: string
        priority:
          type: string
        assigneeIds:
          type: array
          items:
            type: integer
        startDate:
          type: string
        dueDate:
          type: string
        effort:
          type: number
        billable:
          type: boolean
        isMilestone:
          type: boolean

    # ─── USER ───────────────────────────────────────────────────
...
    TimeEntry:
      type: object
      properties:
        id:
          type: integer
        projectId:
          type: integer
        userId:
          type: integer
        taskId:
          type: ["integer", "null"]
        date:
          type: string
        hours:
          type: number
        description:
          type: string
        billable:
          type: boolean
        approved:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - projectId
        - userId
        - date
        - hours
        - description
        - billable
        - approved
        - createdAt
    CreateTimeEntryBody:
      type: object
      properties:
        projectId:
          type: integer
        userId:
          type: integer
        taskId:
          type: integer
        date:
          type: string
        hours:
          type: number
        description:
          type: string
        billable:
          type: boolean
      required:
        - projectId
        - userId
        - date
        - hours
        - description
        - billable
    UpdateTimeEntryBody:
      type: object
      properties:
        hours:
          type: number
        description:
          type: string
        billable:
          type: boolean
        approved:
          type: boolean
...
    Timesheet:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        weekStart:
          type: string
        status:
          type: string
        totalHours:
          type: number
        billableHours:
          type: number
        submittedAt:
          type: ["string", "null"]
        approvedAt:
          type: ["string", "null"]
        approvedByUserId:
          type: ["integer", "null"]
        rejectionNote:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - userId
        - weekStart
        - status
        - totalHours
        - billableHours
        - createdAt
    CreateTimesheetBody:
      type: object
      properties:
        userId:
          type: integer
        weekStart:
          type: string
      required:
        - userId
        - weekStart
    UpdateTimesheetBody:
      type: object
      properties:
        totalHours:
          type: number
        billableHours:
          type: number

    # ─── INVOICE LINE ITEM ───────────────────────────────────────
...